Create scr_server
Modification of scr_server from original scr_patch in torcs. Now works in TCP-IP for easier communication of big data like the image. Warning : It is not necessarily sending the camera of the bot, unfortunately it just reads the pixels from the UI! (if anyone knows how to do that please let me know! :-) )